+++ src/index.js
@@ -11,7 +11,7 @@ export const ReactTableDefaults = defaultProps
 export default class ReactTable extends Methods(Lifecycle(Component)) {
   static defaultProps = defaultProps
 
-  constructor (props) {
+  constructor(props) {
     super()
 
     this.getResolvedState = this.getResolvedState.bind(this)
@@ -43,7 +43,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     }
   }
 
-  render () {
+  render() {
     const resolvedState = this.getResolvedState()
     const {
       children,
@@ -148,7 +148,11 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
           }
           const newPath = path.concat([i])
           if (rowWithViewIndex[subRowsKey] && _.get(expanded, newPath)) {
-            [rowWithViewIndex[subRowsKey], index] = recurseRowsViewIndex(rowWithViewIndex[subRowsKey], newPath, index)
+            ;[rowWithViewIndex[subRowsKey], index] = recurseRowsViewIndex(
+              rowWithViewIndex[subRowsKey],
+              newPath,
+              index
+            )
           }
           return rowWithViewIndex
         }),
@@ -156,15 +160,17 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       ]
     }
 
-    [pageRows] = recurseRowsViewIndex(pageRows)
+    ;[pageRows] = recurseRowsViewIndex(pageRows)
 
     const canPrevious = page > 0
     const canNext = page + 1 < pages
 
-    const rowMinWidth = _.sum(allVisibleColumns.map(d => {
-      const resizedColumn = resized.find(x => x.id === d.id) || {}
-      return _.getFirstDefined(resizedColumn.value, d.width, d.minWidth)
-    }))
+    const rowMinWidth = _.sum(
+      allVisibleColumns.map(d => {
+        const resizedColumn = resized.find(x => x.id === d.id) || {}
+        return _.getFirstDefined(resizedColumn.value, d.width, d.minWidth)
+      })
+    )
 
     let rowIndex = -1
 
@@ -184,8 +190,12 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     // Visual Components
 
     const makeHeaderGroups = () => {
-      const theadGroupProps = _.splitProps(getTheadGroupProps(finalState, undefined, undefined, this))
-      const theadGroupTrProps = _.splitProps(getTheadGroupTrProps(finalState, undefined, undefined, this))
+      const theadGroupProps = _.splitProps(
+        getTheadGroupProps(finalState, undefined, undefined, this)
+      )
+      const theadGroupTrProps = _.splitProps(
+        getTheadGroupTrProps(finalState, undefined, undefined, this)
+      )
       return (
         <TheadComponent
           className={classnames('-headerGroups', theadGroupProps.className)}
@@ -207,13 +217,30 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     }
 
     const makeHeaderGroup = (column, i) => {
-      const resizedValue = col => (resized.find(x => x.id === col.id) || {}).value
-      const flex = _.sum(column.columns.map(col => col.width || resizedValue(col) ? 0 : col.minWidth))
-      const width = _.sum(column.columns.map(col => _.getFirstDefined(resizedValue(col), col.width, col.minWidth)))
-      const maxWidth = _.sum(column.columns.map(col => _.getFirstDefined(resizedValue(col), col.width, col.maxWidth)))
+      const resizedValue = col =>
+        (resized.find(x => x.id === col.id) || {}).value
+      const flex = _.sum(
+        column.columns.map(
+          col => (col.width || resizedValue(col) ? 0 : col.minWidth)
+        )
+      )
+      const width = _.sum(
+        column.columns.map(col =>
+          _.getFirstDefined(resizedValue(col), col.width, col.minWidth)
+        )
+      )
+      const maxWidth = _.sum(
+        column.columns.map(col =>
+          _.getFirstDefined(resizedValue(col), col.width, col.maxWidth)
+        )
+      )
 
-      const theadGroupThProps = _.splitProps(getTheadGroupThProps(finalState, undefined, column, this))
-      const columnHeaderProps = _.splitProps(column.getHeaderProps(finalState, undefined, column, this))
+      const theadGroupThProps = _.splitProps(
+        getTheadGroupThProps(finalState, undefined, column, this)
+      )
+      const columnHeaderProps = _.splitProps(
+        column.getHeaderProps(finalState, undefined, column, this)
+      )
 
       const classes = [
         column.headerClassName,
@@ -241,9 +268,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       return (
         <ThComponent
           key={i + '-' + column.id}
-          className={classnames(
-            classes
-          )}
+          className={classnames(classes)}
           style={{
             ...styles,
             ...flexStyles
@@ -259,8 +284,12 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     }
 
     const makeHeaders = () => {
-      const theadProps = _.splitProps(getTheadProps(finalState, undefined, undefined, this))
-      const theadTrProps = _.splitProps(getTheadTrProps(finalState, undefined, undefined, this))
+      const theadProps = _.splitProps(
+        getTheadProps(finalState, undefined, undefined, this)
+      )
+      const theadTrProps = _.splitProps(
+        getTheadTrProps(finalState, undefined, undefined, this)
+      )
       return (
         <TheadComponent
           className={classnames('-header', theadProps.className)}
@@ -284,11 +313,25 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     const makeHeader = (column, i) => {
       const resizedCol = resized.find(x => x.id === column.id) || {}
       const sort = sorted.find(d => d.id === column.id)
-      const show = typeof column.show === 'function' ? column.show() : column.show
-      const width = _.getFirstDefined(resizedCol.value, column.width, column.minWidth)
-      const maxWidth = _.getFirstDefined(resizedCol.value, column.width, column.maxWidth)
-      const theadThProps = _.splitProps(getTheadThProps(finalState, undefined, column, this))
-      const columnHeaderProps = _.splitProps(column.getHeaderProps(finalState, undefined, column, this))
+      const show = typeof column.show === 'function'
+        ? column.show()
+        : column.show
+      const width = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.minWidth
+      )
+      const maxWidth = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.maxWidth
+      )
+      const theadThProps = _.splitProps(
+        getTheadThProps(finalState, undefined, column, this)
+      )
+      const columnHeaderProps = _.splitProps(
+        column.getHeaderProps(finalState, undefined, column, this)
+      )
 
       const classes = [
         column.headerClassName,
@@ -308,14 +351,13 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       }
 
       const isResizable = _.getFirstDefined(column.resizable, resizable, false)
-      const resizer = isResizable ? (
-        <ResizerComponent
-          onMouseDown={e => this.resizeColumnStart(column, e, false)}
-          onTouchStart={e => this.resizeColumnStart(column, e, true)}
-
-          {...resizerProps}
-        />
-      ) : null
+      const resizer = isResizable
+        ? <ResizerComponent
+            onMouseDown={e => this.resizeColumnStart(column, e, false)}
+            onTouchStart={e => this.resizeColumnStart(column, e, true)}
+            {...resizerProps}
+          />
+        : null
 
       const isSortable = _.getFirstDefined(column.sortable, sortable, false)
 
@@ -328,7 +370,9 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
             sort ? (sort.desc ? '-sort-desc' : '-sort-asc') : '',
             isSortable && '-cursor-pointer',
             !show && '-hidden',
-            pivotBy && pivotBy.slice(0, -1).includes(column.id) && 'rt-header-pivot'
+            pivotBy &&
+              pivotBy.slice(0, -1).includes(column.id) &&
+              'rt-header-pivot'
           )}
           style={{
             ...styles,
@@ -336,12 +380,12 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
             width: `${width}px`,
             maxWidth: `${maxWidth}px`
           }}
-          toggleSort={(e) => {
+          toggleSort={e => {
             isSortable && this.sortColumn(column, e.shiftKey)
           }}
           {...rest}
         >
-          <div className='rt-resizable-header-content'>
+          <div className="rt-resizable-header-content">
             {_.normalizeComponent(column.Header, {
               data: sortedData,
               column: column
@@ -353,8 +397,12 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     }
 
     const makeFilters = () => {
-      const theadFilterProps = _.splitProps(getTheadFilterProps(finalState, undefined, undefined, this))
-      const theadFilterTrProps = _.splitProps(getTheadFilterTrProps(finalState, undefined, undefined, this))
+      const theadFilterProps = _.splitProps(
+        getTheadFilterProps(finalState, undefined, undefined, this)
+      )
+      const theadFilterTrProps = _.splitProps(
+        getTheadFilterTrProps(finalState, undefined, undefined, this)
+      )
       return (
         <TheadComponent
           className={classnames('-filters', theadFilterProps.className)}
@@ -377,10 +425,22 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
 
     const makeFilter = (column, i) => {
       const resizedCol = resized.find(x => x.id === column.id) || {}
-      const width = _.getFirstDefined(resizedCol.value, column.width, column.minWidth)
-      const maxWidth = _.getFirstDefined(resizedCol.value, column.width, column.maxWidth)
-      const theadFilterThProps = _.splitProps(getTheadFilterThProps(finalState, undefined, column, this))
-      const columnHeaderProps = _.splitProps(column.getHeaderProps(finalState, undefined, column, this))
+      const width = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.minWidth
+      )
+      const maxWidth = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.maxWidth
+      )
+      const theadFilterThProps = _.splitProps(
+        getTheadFilterThProps(finalState, undefined, column, this)
+      )
+      const columnHeaderProps = _.splitProps(
+        column.getHeaderProps(finalState, undefined, column, this)
+      )
 
       const classes = [
         column.headerClassName,
@@ -403,14 +463,16 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
 
       const ResolvedFilterComponent = column.Filter || FilterComponent
 
-      const isFilterable = _.getFirstDefined(column.filterable, filterable, false)
+      const isFilterable = _.getFirstDefined(
+        column.filterable,
+        filterable,
+        false
+      )
 
       return (
         <ThComponent
           key={i + '-' + column.id}
-          className={classnames(
-            classes
-          )}
+          className={classnames(classes)}
           style={{
             ...styles,
             flex: `${width} 0 auto`,
@@ -419,16 +481,17 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
           }}
           {...rest}
         >
-          {isFilterable ? (
-            _.normalizeComponent(ResolvedFilterComponent,
-              {
-                column,
-                filter,
-                onChange: (value) => (this.filterColumn(column, value))
-              },
-              defaultProps.column.Filter
-            )
-          ) : null}
+          {isFilterable
+            ? _.normalizeComponent(
+                ResolvedFilterComponent,
+                {
+                  column,
+                  filter,
+                  onChange: value => this.filterColumn(column, value)
+                },
+                defaultProps.column.Filter
+              )
+            : null}
         </ThComponent>
       )
     }
@@ -447,12 +510,11 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       }
       const isExpanded = _.get(expanded, rowInfo.nestingPath)
       const trGroupProps = getTrGroupProps(finalState, rowInfo, undefined, this)
-      const trProps = _.splitProps(getTrProps(finalState, rowInfo, undefined, this))
+      const trProps = _.splitProps(
+        getTrProps(finalState, rowInfo, undefined, this)
+      )
       return (
-        <TrGroupComponent
-          key={rowInfo.nestingPath.join('_')}
-          {...trGroupProps}
-        >
+        <TrGroupComponent key={rowInfo.nestingPath.join('_')} {...trGroupProps}>
           <TrComponent
             className={classnames(
               trProps.className,
@@ -463,11 +525,25 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
           >
             {allVisibleColumns.map((column, i2) => {
               const resizedCol = resized.find(x => x.id === column.id) || {}
-              const show = typeof column.show === 'function' ? column.show() : column.show
-              const width = _.getFirstDefined(resizedCol.value, column.width, column.minWidth)
-              const maxWidth = _.getFirstDefined(resizedCol.value, column.width, column.maxWidth)
-              const tdProps = _.splitProps(getTdProps(finalState, rowInfo, column, this))
-              const columnProps = _.splitProps(column.getProps(finalState, rowInfo, column, this))
+              const show = typeof column.show === 'function'
+                ? column.show()
+                : column.show
+              const width = _.getFirstDefined(
+                resizedCol.value,
+                column.width,
+                column.minWidth
+              )
+              const maxWidth = _.getFirstDefined(
+                resizedCol.value,
+                column.width,
+                column.maxWidth
+              )
+              const tdProps = _.splitProps(
+                getTdProps(finalState, rowInfo, column, this)
+              )
+              const columnProps = _.splitProps(
+                column.getProps(finalState, rowInfo, column, this)
+              )
 
               const classes = [
                 tdProps.className,
@@ -484,7 +560,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
               const cellInfo = {
                 ...rowInfo,
                 isExpanded,
-                column: {...column},
+                column: { ...column },
                 value: rowInfo.row[column.id],
                 pivoted: column.pivoted,
                 expander: column.expander,
@@ -504,7 +580,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
               let isBranch
               let isPreview
 
-              const onExpanderClick = (e) => {
+              const onExpanderClick = e => {
                 let newExpanded = _.clone(expanded)
                 if (isExpanded) {
                   newExpanded = _.set(newExpanded, cellInfo.nestingPath, false)
@@ -512,29 +588,41 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
                   newExpanded = _.set(newExpanded, cellInfo.nestingPath, {})
                 }
 
-                return this.setStateWithData({
-                  expanded: newExpanded
-                }, () => {
-                  onExpandedChange && onExpandedChange(newExpanded, cellInfo.nestingPath, e)
-                })
+                return this.setStateWithData(
+                  {
+                    expanded: newExpanded
+                  },
+                  () => {
+                    onExpandedChange &&
+                      onExpandedChange(newExpanded, cellInfo.nestingPath, e)
+                  }
+                )
               }
 
               // Default to a standard cell
-              let resolvedCell = _.normalizeComponent(column.Cell, cellInfo, value)
+              let resolvedCell = _.normalizeComponent(
+                column.Cell,
+                cellInfo,
+                value
+              )
 
               // Resolve Renderers
-              const ResolvedAggregatedComponent = column.Aggregated || (!column.aggregate ? AggregatedComponent : column.Cell)
-              const ResolvedExpanderComponent = column.Expander || ExpanderComponent
-              const ResolvedPivotValueComponent = column.PivotValue || PivotValueComponent
-              const DefaultResolvedPivotComponent = PivotComponent || (
-                  props => (
-                    <div>
-                      <ResolvedExpanderComponent {...props} />
-                      <ResolvedPivotValueComponent {...props} />
-                    </div>
-                  )
-                )
-              const ResolvedPivotComponent = column.Pivot || DefaultResolvedPivotComponent
+              const ResolvedAggregatedComponent =
+                column.Aggregated ||
+                (!column.aggregate ? AggregatedComponent : column.Cell)
+              const ResolvedExpanderComponent =
+                column.Expander || ExpanderComponent
+              const ResolvedPivotValueComponent =
+                column.PivotValue || PivotValueComponent
+              const DefaultResolvedPivotComponent =
+                PivotComponent ||
+                (props =>
+                  <div>
+                    <ResolvedExpanderComponent {...props} />
+                    <ResolvedPivotValueComponent {...props} />
+                  </div>)
+              const ResolvedPivotComponent =
+                column.Pivot || DefaultResolvedPivotComponent
 
               // Is this cell expandable?
               if (cellInfo.pivoted || cellInfo.expander) {
@@ -556,30 +644,47 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
 
               if (cellInfo.pivoted) {
                 // Is this column a branch?
-                isBranch = rowInfo.row[pivotIDKey] === column.id &&
-                  cellInfo.subRows
+                isBranch =
+                  rowInfo.row[pivotIDKey] === column.id && cellInfo.subRows
                 // Should this column be blank?
-                isPreview = pivotBy.indexOf(column.id) > pivotBy.indexOf(rowInfo.row[pivotIDKey]) &&
-                  cellInfo.subRows
+                isPreview =
+                  pivotBy.indexOf(column.id) >
+                    pivotBy.indexOf(rowInfo.row[pivotIDKey]) && cellInfo.subRows
                 // Pivot Cell Render Override
                 if (isBranch) {
                   // isPivot
-                  resolvedCell = _.normalizeComponent(ResolvedPivotComponent, {
-                    ...cellInfo,
-                    value: row[pivotValKey]
-                  }, row[pivotValKey])
+                  resolvedCell = _.normalizeComponent(
+                    ResolvedPivotComponent,
+                    {
+                      ...cellInfo,
+                      value: row[pivotValKey]
+                    },
+                    row[pivotValKey]
+                  )
                 } else if (isPreview) {
                   // Show the pivot preview
-                  resolvedCell = _.normalizeComponent(ResolvedAggregatedComponent, cellInfo, value)
+                  resolvedCell = _.normalizeComponent(
+                    ResolvedAggregatedComponent,
+                    cellInfo,
+                    value
+                  )
                 } else {
                   resolvedCell = null
                 }
               } else if (cellInfo.aggregated) {
-                resolvedCell = _.normalizeComponent(ResolvedAggregatedComponent, cellInfo, value)
+                resolvedCell = _.normalizeComponent(
+                  ResolvedAggregatedComponent,
+                  cellInfo,
+                  value
+                )
               }
 
               if (cellInfo.expander) {
-                resolvedCell = _.normalizeComponent(ResolvedExpanderComponent, cellInfo, row[pivotValKey])
+                resolvedCell = _.normalizeComponent(
+                  ResolvedExpanderComponent,
+                  cellInfo,
+                  row[pivotValKey]
+                )
                 if (pivotBy) {
                   if (cellInfo.groupedByPivot) {
                     resolvedCell = null
@@ -614,24 +719,31 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
               )
             })}
           </TrComponent>
-          {(
-            rowInfo.subRows &&
+          {rowInfo.subRows &&
             isExpanded &&
-            rowInfo.subRows.map((d, i) => makePageRow(d, i, rowInfo.nestingPath))
-          )}
-          {SubComponent && !rowInfo.subRows && isExpanded && SubComponent(rowInfo)}
+            rowInfo.subRows.map((d, i) =>
+              makePageRow(d, i, rowInfo.nestingPath)
+            )}
+          {SubComponent &&
+            !rowInfo.subRows &&
+            isExpanded &&
+            SubComponent(rowInfo)}
         </TrGroupComponent>
       )
     }
 
     const makePadRow = (row, i) => {
-      const trGroupProps = getTrGroupProps(finalState, undefined, undefined, this)
-      const trProps = _.splitProps(getTrProps(finalState, undefined, undefined, this))
+      const trGroupProps = getTrGroupProps(
+        finalState,
+        undefined,
+        undefined,
+        this
+      )
+      const trProps = _.splitProps(
+        getTrProps(finalState, undefined, undefined, this)
+      )
       return (
-        <TrGroupComponent
-          key={i}
-          {...trGroupProps}
-        >
+        <TrGroupComponent key={i} {...trGroupProps}>
           <TrComponent
             className={classnames(
               '-padRow',
@@ -648,12 +760,26 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
 
     const makePadColumn = (column, i) => {
       const resizedCol = resized.find(x => x.id === column.id) || {}
-      const show = typeof column.show === 'function' ? column.show() : column.show
-      let width = _.getFirstDefined(resizedCol.value, column.width, column.minWidth)
+      const show = typeof column.show === 'function'
+        ? column.show()
+        : column.show
+      let width = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.minWidth
+      )
       let flex = width
-      let maxWidth = _.getFirstDefined(resizedCol.value, column.width, column.maxWidth)
-      const tdProps = _.splitProps(getTdProps(finalState, undefined, column, this))
-      const columnProps = _.splitProps(column.getProps(finalState, undefined, column, this))
+      let maxWidth = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.maxWidth
+      )
+      const tdProps = _.splitProps(
+        getTdProps(finalState, undefined, column, this)
+      )
+      const columnProps = _.splitProps(
+        column.getProps(finalState, undefined, column, this)
+      )
 
       const classes = [
         tdProps.className,
@@ -670,10 +796,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       return (
         <TdComponent
           key={i + '-' + column.id}
-          className={classnames(
-            classes,
-            !show && 'hidden'
-          )}
+          className={classnames(classes, !show && 'hidden')}
           style={{
             ...styles,
             flex: `${flex} 0 auto`,
@@ -681,15 +804,15 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
             maxWidth: `${maxWidth}px`
           }}
           {...tdProps.rest}
-        >
-          &nbsp;
-        </TdComponent>
+        />
       )
     }
 
     const makeColumnFooters = () => {
       const tFootProps = getTfootProps(finalState, undefined, undefined, this)
-      const tFootTrProps = _.splitProps(getTfootTrProps(finalState, undefined, undefined, this))
+      const tFootTrProps = _.splitProps(
+        getTfootTrProps(finalState, undefined, undefined, this)
+      )
       return (
         <TfootComponent
           className={tFootProps.className}
@@ -700,9 +823,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
           {...tFootProps.rest}
         >
           <TrComponent
-            className={classnames(
-              tFootTrProps.className
-            )}
+            className={classnames(tFootTrProps.className)}
             style={tFootTrProps.style}
             {...tFootTrProps.rest}
           >
@@ -714,12 +835,28 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
 
     const makeColumnFooter = (column, i) => {
       const resizedCol = resized.find(x => x.id === column.id) || {}
-      const show = typeof column.show === 'function' ? column.show() : column.show
-      const width = _.getFirstDefined(resizedCol.value, column.width, column.minWidth)
-      const maxWidth = _.getFirstDefined(resizedCol.value, column.width, column.maxWidth)
-      const tFootTdProps = _.splitProps(getTfootTdProps(finalState, undefined, undefined, this))
-      const columnProps = _.splitProps(column.getProps(finalState, undefined, column, this))
-      const columnFooterProps = _.splitProps(column.getFooterProps(finalState, undefined, column, this))
+      const show = typeof column.show === 'function'
+        ? column.show()
+        : column.show
+      const width = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.minWidth
+      )
+      const maxWidth = _.getFirstDefined(
+        resizedCol.value,
+        column.width,
+        column.maxWidth
+      )
+      const tFootTdProps = _.splitProps(
+        getTfootTdProps(finalState, undefined, undefined, this)
+      )
+      const columnProps = _.splitProps(
+        column.getProps(finalState, undefined, column, this)
+      )
+      const columnFooterProps = _.splitProps(
+        column.getFooterProps(finalState, undefined, column, this)
+      )
 
       const classes = [
         tFootTdProps.className,
@@ -738,10 +875,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       return (
         <TdComponent
           key={i + '-' + column.id}
-          className={classnames(
-            classes,
-            !show && 'hidden'
-          )}
+          className={classnames(classes, !show && 'hidden')}
           style={{
             ...styles,
             flex: `${width} 0 auto`,
@@ -761,23 +895,33 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
     }
 
     const makePagination = () => {
-      const paginationProps = _.splitProps(getPaginationProps(finalState, undefined, undefined, this))
-      return <PaginationComponent
-        {...resolvedState}
-        pages={pages}
-        canPrevious={canPrevious}
-        canNext={canNext}
-        onPageChange={this.onPageChange}
-        onPageSizeChange={this.onPageSizeChange}
-        className={paginationProps.className}
-        style={paginationProps.style}
-        {...paginationProps.rest}
-      />
+      const paginationProps = _.splitProps(
+        getPaginationProps(finalState, undefined, undefined, this)
+      )
+      return (
+        <PaginationComponent
+          {...resolvedState}
+          pages={pages}
+          canPrevious={canPrevious}
+          canNext={canNext}
+          onPageChange={this.onPageChange}
+          onPageSizeChange={this.onPageSizeChange}
+          className={paginationProps.className}
+          style={paginationProps.style}
+          {...paginationProps.rest}
+        />
+      )
     }
 
-    const rootProps = _.splitProps(getProps(finalState, undefined, undefined, this))
-    const tableProps = _.splitProps(getTableProps(finalState, undefined, undefined, this))
-    const tBodyProps = _.splitProps(getTbodyProps(finalState, undefined, undefined, this))
+    const rootProps = _.splitProps(
+      getProps(finalState, undefined, undefined, this)
+    )
+    const tableProps = _.splitProps(
+      getTableProps(finalState, undefined, undefined, this)
+    )
+    const tBodyProps = _.splitProps(
+      getTbodyProps(finalState, undefined, undefined, this)
+    )
     const loadingProps = getLoadingProps(finalState, undefined, undefined, this)
     const noDataProps = getNoDataProps(finalState, undefined, undefined, this)
     const resizerProps = getResizerProps(finalState, undefined, undefined, this)
@@ -786,11 +930,7 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
       const pagination = makePagination()
       return (
         <div
-          className={classnames(
-            'ReactTable',
-            className,
-            rootProps.className
-          )}
+          className={classnames('ReactTable', className, rootProps.className)}
           style={{
             ...style,
             ...rootProps.style
@@ -798,9 +938,9 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
           {...rootProps.rest}
         >
           {showPagination && showPaginationTop
-            ? <div className='pagination-top'>
-              {pagination}
-            </div>
+            ? <div className="pagination-top">
+                {pagination}
+              </div>
             : null}
           <TableComponent
             className={classnames(
@@ -827,17 +967,14 @@ export default class ReactTable extends Methods(Lifecycle(Component)) {
             {hasColumnFooter ? makeColumnFooters() : null}
           </TableComponent>
           {showPagination && showPaginationBottom
-            ? <div className='pagination-bottom'>
-              {pagination}
-            </div>
+            ? <div className="pagination-bottom">
+                {pagination}
+              </div>
             : null}
-          {!pageRows.length && (
-            <NoDataComponent
-              {...noDataProps}
-            >
+          {!pageRows.length &&
+            <NoDataComponent {...noDataProps}>
               {_.normalizeComponent(noDataText)}
-            </NoDataComponent>
-          )}
+            </NoDataComponent>}
           <LoadingComponent
             loading={loading}
             loadingText={loadingText}