+++ src/pagination.js
@@ -3,9 +3,8 @@ import classnames from 'classnames'
 //
 // import _ from './utils'
 
-const defaultButton = (props) => (
+const defaultButton = props =>
   <button type='button' {...props} className='-btn'>{props.children}</button>
-)
 
 export default class ReactTablePagination extends Component {
   constructor (props) {
@@ -16,12 +15,12 @@ export default class ReactTablePagination extends Component {
     this.applyPage = this.applyPage.bind(this)
 
     this.state = {
-      page: props.page
+      page: props.page,
     }
   }
 
   componentWillReceiveProps (nextProps) {
-    this.setState({page: nextProps.page})
+    this.setState({ page: nextProps.page })
   }
 
   getSafePage (page) {
@@ -33,7 +32,7 @@ export default class ReactTablePagination extends Component {
 
   changePage (page) {
     page = this.getSafePage(page)
-    this.setState({page})
+    this.setState({ page })
     if (this.props.page !== page) {
       this.props.onPageChange(page)
     }
@@ -60,7 +59,7 @@ export default class ReactTablePagination extends Component {
       onPageSizeChange,
       className,
       PreviousComponent = defaultButton,
-      NextComponent = defaultButton
+      NextComponent = defaultButton,
     } = this.props
 
     return (
@@ -70,7 +69,7 @@ export default class ReactTablePagination extends Component {
       >
         <div className='-previous'>
           <PreviousComponent
-            onClick={(e) => {
+            onClick={e => {
               if (!canPrevious) return
               this.changePage(page - 1)
             }}
@@ -81,53 +80,51 @@ export default class ReactTablePagination extends Component {
         </div>
         <div className='-center'>
           <span className='-pageInfo'>
-            {this.props.pageText} {showPageJump ? (
-              <div className='-pageJump'>
-                <input
-                  type={this.state.page === '' ? 'text' : 'number'}
-                  onChange={e => {
-                    const val = e.target.value
-                    const page = val - 1
-                    if (val === '') {
-                      return this.setState({page: val})
-                    }
-                    this.setState({page: this.getSafePage(page)})
-                  }}
-                  value={this.state.page === '' ? '' : this.state.page + 1}
-                  onBlur={this.applyPage}
-                  onKeyPress={e => {
-                    if (e.which === 13 || e.keyCode === 13) {
-                      this.applyPage()
-                    }
-                  }}
-                />
-              </div>
-            ) : (
-              <span className='-currentPage'>{page + 1}</span>
-            )} {this.props.ofText} <span className='-totalPages'>{pages || 1}</span>
+            {this.props.pageText}{' '}
+            {showPageJump
+              ? <div className='-pageJump'>
+                  <input
+                    type={this.state.page === '' ? 'text' : 'number'}
+                    onChange={e => {
+                      const val = e.target.value
+                      const page = val - 1
+                      if (val === '') {
+                        return this.setState({ page: val })
+                      }
+                      this.setState({ page: this.getSafePage(page) })
+                    }}
+                    value={this.state.page === '' ? '' : this.state.page + 1}
+                    onBlur={this.applyPage}
+                    onKeyPress={e => {
+                      if (e.which === 13 || e.keyCode === 13) {
+                        this.applyPage()
+                      }
+                    }}
+                  />
+                </div>
+              : <span className='-currentPage'>{page + 1}</span>}{' '}
+            {this.props.ofText}{' '}
+            <span className='-totalPages'>{pages || 1}</span>
           </span>
-          {showPageSizeOptions && (
+          {showPageSizeOptions &&
             <span className='select-wrap -pageSizeOptions'>
               <select
-                onChange={(e) => onPageSizeChange(Number(e.target.value))}
+                onChange={e => onPageSizeChange(Number(e.target.value))}
                 value={pageSize}
               >
                 {pageSizeOptions.map((option, i) => {
                   return (
-                    <option
-                      key={i}
-                      value={option}>
+                    <option key={i} value={option}>
                       {option} {this.props.rowsText}
                     </option>
                   )
                 })}
               </select>
-            </span>
-          )}
+            </span>}
         </div>
         <div className='-next'>
           <NextComponent
-            onClick={(e) => {
+            onClick={e => {
               if (!canNext) return
               this.changePage(page + 1)
             }}